feature(core): update response message

+++ src/productos/productos.controller.ts
@@ -7,18 +7,18 @@ import {
   Put, 
   Delete, 
   NotFoundException, 
-  ConflictException, 
   BadRequestException 
 } from '@nestjs/common';
 import { ProductosService } from './productos.service';
+import { TiendasService } from 'src/tiendas/tiendas.service';
 import { CreateProductoDto } from './dto/create-producto.dto';
 import { UpdateProductoDto } from './dto/update-producto.dto';
 import { ObjectId } from 'mongodb';  
-import { Types } from 'mongoose';
+import { Types,isValidObjectId } from 'mongoose';
 
 @Controller('products')
 export class ProductosController {
-  constructor(private readonly productosService: ProductosService) {}
+  constructor(private readonly productosService: ProductosService, private readonly tiendasService: TiendasService,) {}
 
   @Post(':productId/stores/:storeId')
   async addStoreToProduct(
@@ -26,32 +26,58 @@ export class ProductosController {
     @Param('storeId') storeId: string,
   ) {
     try {
+      // Verificar que el producto existe
+      const productExists = await this.productosService.findOne(productId);
+      if (!productExists) {
+        throw new NotFoundException(`Producto con ID ${productId} no encontrado`);
+      }
+  
+      // Verificar que la tienda existe
+      const storeExists = await this.tiendasService.findOne(storeId);
+      if (!storeExists) {
+        throw new NotFoundException(`Tienda con ID ${storeId} no encontrada`);
+      }
+  
+      // Agregar la tienda al producto si ambos existen
       const result = await this.productosService.addStoreToProduct(productId, storeId);
       
       return { message: 'Tienda añadida al producto con éxito', result };
     } catch (error) {
-      return { 
-        error: 'Hubo un error al añadir la tienda al producto', 
-        details: error.message 
+      return {
+        error: 'Hubo un error al añadir la tienda al producto',
+        details: error.message || error,
       };
     }
   }
+  
 
+ 
   @Get(':productId/stores')
   async findStoresFromProduct(@Param('productId') productId: string) {
     try {
-      const stores = await this.productosService.findStoresFromProduct(productId);
-      if (!stores.length) {
-        throw new NotFoundException('No se encontraron tiendas asociadas a este producto');
+       if (!isValidObjectId(productId)) {
+        throw new BadRequestException(
+          `El ID proporcionado (${productId}) no es válido`,
+        );
       }
+  
+       const stores = await this.productosService.findStoresFromProduct(productId);
+  
+       if (!stores.length) {
+        throw new NotFoundException(
+          'No se encontraron tiendas asociadas a este producto',
+        );
+      }
+  
       return stores;
     } catch (error) {
-      return { 
-        error: 'Hubo un error al obtener las tiendas asociadas al producto', 
-        details: error.message 
+      return {
+        error: 'Hubo un error al obtener las tiendas asociadas al producto',
+        details: error.message,
       };
     }
   }
+  
 
   @Get(':productId/stores/:storeId')
   async findStoreFromProduct(
@@ -59,18 +85,28 @@ export class ProductosController {
     @Param('storeId') storeId: string,
   ) {
     try {
-      const store = await this.productosService.findStoreFromProduct(productId, storeId);
-      if (!store) {
+       if (!Types.ObjectId.isValid(productId) || !Types.ObjectId.isValid(storeId)) {
+        throw new BadRequestException('Uno o ambos IDs no son válidos');
+      }
+
+       const store = await this.productosService.findStoreFromProduct(productId, storeId);
+
+       if (!store) {
         throw new NotFoundException('Tienda no encontrada para este producto');
       }
-      return store;
+
+       return { 
+        message: 'Tienda encontrada con éxito', 
+        store 
+      };
     } catch (error) {
-      return { 
+       return { 
         error: 'Hubo un error al obtener la tienda asociada al producto', 
-        details: error.message 
+        details: error.message || error.toString() 
       };
     }
   }
+ 
 
   @Put(':productId/stores')
   async updateStoresFromProduct(
@@ -94,12 +130,22 @@ export class ProductosController {
     @Param('storeId') storeId: string,
   ) {
     try {
-      const result = await this.productosService.deleteStoreFromProduct(productId, storeId);
-       return { message: 'Tienda eliminada del producto con éxito', result };
+      
+      if (!Types.ObjectId.isValid(productId) || !Types.ObjectId.isValid(storeId)) {
+        throw new BadRequestException('Uno o ambos IDs no son válidos');
+      }
+
+       const result = await this.productosService.deleteStoreFromProduct(productId, storeId);
+      
+      return { 
+        message: 'Tienda eliminada del producto con éxito', 
+        result 
+      };
     } catch (error) {
+      
       return { 
         error: 'Hubo un error al eliminar la tienda del producto', 
-        details: error.message 
+        details: error.message || error.toString() 
       };
     }
   }
